requirements.txt
kollavarsham
multipledispatch
pillow
pytz
urllib3<2.0
requests
